further VideoWorkspace cleanup
