+++ packages/atlas/src/components/_inputs/MultiFileSelect/MultiFileSelect.stories.tsx
@@ -21,7 +21,6 @@ export default {
 } as Meta
 
 const Template: Story<MultiFileSelectProps> = (args) => {
-  const [error, setError] = useState<string | null>(null)
   const [files, setFiles] = useState<InputFilesState>({
     video: null,
     thumbnail: null,
@@ -31,8 +30,6 @@ const Template: Story<MultiFileSelectProps> = (args) => {
     <MultiFileSelect
       {...args}
       files={files}
-      error={error}
-      onError={setError}
       onThumbnailChange={(thumbnail) => setFiles((files) => ({ ...files, thumbnail }))}
       onVideoChange={(video) => setFiles((files) => ({ ...files, video }))}
     />
