+++ packages/atlas/src/views/studio/VideoWorkspace/VideoForm/index.ts
@@ -1,0 +1,1 @@
+export * from './VideoForm'
