--- packages/atlas/src/views/studio/VideoWorkspace/VideoWorkspaceForm/index.ts
@@ -1,1 +1,0 @@
-export * from './VideoWorkspaceForm'
